Refactor TOC HTML/CSS
Fixes #2

This commit introduces a new data attribute added to a top level
html node. This attribute is a flag which determines the open/close
state of the TOC. CSS rules are rewritten to refrence this. A CSS variable
is introduced for controlling TOC width.

The top level container is no longer a grid.

The TOC is rendered offscreen. When it is toggled to display, all of the
page content is moved using transforms sideways, revealing the TOC.
Because no page elements change size, no reflow occurs, solving
the problem in #2.

The close button on the TOC could often scroll out of view. To fix this,
the TOC HTML has been tweaked so that a header with the title and close
button stays fixed, and only the TOC UL scrolls.